#233 Further changes on model factories (TopicController)
