+++ src/Presentation/Nop.Web/Controllers/TopicController.cs
@@ -1,18 +1,10 @@
-﻿using System;
-using System.Linq;
-using System.Web.Mvc;
-using Nop.Core;
-using Nop.Core.Caching;
-using Nop.Core.Domain.Topics;
-using Nop.Services.Customers;
+﻿using System.Web.Mvc;
 using Nop.Services.Localization;
 using Nop.Services.Security;
-using Nop.Services.Seo;
 using Nop.Services.Stores;
 using Nop.Services.Topics;
+using Nop.Web.Factories;
 using Nop.Web.Framework.Security;
-using Nop.Web.Infrastructure.Cache;
-using Nop.Web.Models.Topics;
 
 namespace Nop.Web.Controllers
 {
@@ -20,66 +12,30 @@ namespace Nop.Web.Controllers
     {
         #region Fields
 
+        private readonly ITopicModelFactory _topicModelFactory;
         private readonly ITopicService _topicService;
-        private readonly IWorkContext _workContext;
-        private readonly IStoreContext _storeContext;
         private readonly ILocalizationService _localizationService;
-        private readonly ICacheManager _cacheManager;
         private readonly IStoreMappingService _storeMappingService;
         private readonly IAclService _aclService;
-        private readonly ITopicTemplateService _topicTemplateService;
         private readonly IPermissionService _permissionService;
 
         #endregion
 
         #region Constructors
 
-        public TopicController(ITopicService topicService,
+        public TopicController(ITopicModelFactory topicModelFactory,
+            ITopicService topicService,
             ILocalizationService localizationService,
-            IWorkContext workContext, 
-            IStoreContext storeContext,
-            ICacheManager cacheManager,
             IStoreMappingService storeMappingService,
             IAclService aclService,
-            ITopicTemplateService topicTemplateService,
             IPermissionService permissionService)
         {
+            this._topicModelFactory = topicModelFactory;
             this._topicService = topicService;
-            this._workContext = workContext;
-            this._storeContext = storeContext;
             this._localizationService = localizationService;
-            this._cacheManager = cacheManager;
             this._storeMappingService = storeMappingService;
             this._aclService = aclService;
-            this._topicTemplateService = topicTemplateService;
             this._permissionService = permissionService;
-        }
-
-        #endregion
-
-        #region Utilities
-
-        [NonAction]
-        protected virtual TopicModel PrepareTopicModel(Topic topic)
-        {
-            if (topic == null)
-                throw new ArgumentNullException("topic");
-
-            var model = new TopicModel
-            {
-                Id = topic.Id,
-                SystemName = topic.SystemName,
-                IncludeInSitemap = topic.IncludeInSitemap,
-                IsPasswordProtected = topic.IsPasswordProtected,
-                Title = topic.IsPasswordProtected ? "" : topic.GetLocalized(x => x.Title),
-                Body = topic.IsPasswordProtected ? "" : topic.GetLocalized(x => x.Body),
-                MetaKeywords = topic.GetLocalized(x => x.MetaKeywords),
-                MetaDescription = topic.GetLocalized(x => x.MetaDescription),
-                MetaTitle = topic.GetLocalized(x => x.MetaTitle),
-                SeName = topic.GetSeName(),
-                TopicTemplateId = topic.TopicTemplateId
-            };
-            return model;
         }
 
         #endregion
@@ -89,118 +45,40 @@ namespace Nop.Web.Controllers
         [NopHttpsRequirement(SslRequirement.No)]
         public ActionResult TopicDetails(int topicId)
         {
-            var cacheKey = string.Format(ModelCacheEventConsumer.TOPIC_MODEL_BY_ID_KEY, 
-                topicId, 
-                _workContext.WorkingLanguage.Id,
-                _storeContext.CurrentStore.Id,
-                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
-            var cacheModel = _cacheManager.Get(cacheKey, () =>
-            {
-                var topic = _topicService.GetTopicById(topicId);
-                if (topic == null)
-                    return null;
-                if (!topic.Published)
-                    return null;
-                //Store mapping
-                if (!_storeMappingService.Authorize(topic))
-                    return null;
-                //ACL (access control list)
-                if (!_aclService.Authorize(topic))
-                    return null;
-                return PrepareTopicModel(topic);
-            }
-            );
-
-            if (cacheModel == null)
+            var model = _topicModelFactory.PrepareTopicModelById(topicId);
+            if (model == null)
                 return RedirectToRoute("HomePage");
-
-            //template
-            var templateCacheKey = string.Format(ModelCacheEventConsumer.TOPIC_TEMPLATE_MODEL_KEY, cacheModel.TopicTemplateId);
-            var templateViewPath = _cacheManager.Get(templateCacheKey, () =>
-            {
-                var template = _topicTemplateService.GetTopicTemplateById(cacheModel.TopicTemplateId);
-                if (template == null)
-                    template = _topicTemplateService.GetAllTopicTemplates().FirstOrDefault();
-                if (template == null)
-                    throw new Exception("No default template could be loaded");
-                return template.ViewPath;
-            });
 
             //display "edit" (manage) link
             if (_permissionService.Authorize(StandardPermissionProvider.AccessAdminPanel) && _permissionService.Authorize(StandardPermissionProvider.ManageTopics))
-                DisplayEditLink(Url.Action("Edit", "Topic", new { id = cacheModel.Id, area = "Admin" }));
+                DisplayEditLink(Url.Action("Edit", "Topic", new { id = model.Id, area = "Admin" }));
 
-            return View(templateViewPath, cacheModel);
+            //template
+            var templateViewPath = _topicModelFactory.PrepareTemplateViewPath(model.TopicTemplateId);
+            return View(templateViewPath, model);
         }
 
         public ActionResult TopicDetailsPopup(string systemName)
         {
-            var cacheKey = string.Format(ModelCacheEventConsumer.TOPIC_MODEL_BY_SYSTEMNAME_KEY,
-                systemName,
-                _workContext.WorkingLanguage.Id,
-                _storeContext.CurrentStore.Id,
-                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
-            var cacheModel = _cacheManager.Get(cacheKey, () =>
-            {
-                //load by store
-                var topic = _topicService.GetTopicBySystemName(systemName, _storeContext.CurrentStore.Id);
-                if (topic == null)
-                    return null;
-                if (!topic.Published)
-                    return null;
-                //ACL (access control list)
-                if (!_aclService.Authorize(topic))
-                    return null;
-                return PrepareTopicModel(topic);
-            });
-
-            if (cacheModel == null)
+            var model = _topicModelFactory.PrepareTopicModelBySystemName(systemName);
+            if (model == null)
                 return RedirectToRoute("HomePage");
 
-            //template
-            var templateCacheKey = string.Format(ModelCacheEventConsumer.TOPIC_TEMPLATE_MODEL_KEY, cacheModel.TopicTemplateId);
-            var templateViewPath = _cacheManager.Get(templateCacheKey, () =>
-            {
-                var template = _topicTemplateService.GetTopicTemplateById(cacheModel.TopicTemplateId);
-                if (template == null)
-                    template = _topicTemplateService.GetAllTopicTemplates().FirstOrDefault();
-                if (template == null)
-                    throw new Exception("No default template could be loaded");
-                return template.ViewPath;
-            });
-
             ViewBag.IsPopup = true;
-            return PartialView(templateViewPath, cacheModel);
+
+            //template
+            var templateViewPath = _topicModelFactory.PrepareTemplateViewPath(model.TopicTemplateId);
+            return PartialView(templateViewPath, model);
         }
 
         [ChildActionOnly]
         public ActionResult TopicBlock(string systemName)
         {
-            var cacheKey = string.Format(ModelCacheEventConsumer.TOPIC_MODEL_BY_SYSTEMNAME_KEY,
-                systemName,
-                _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id,
-                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
-            var cacheModel = _cacheManager.Get(cacheKey, () =>
-            {
-                //load by store
-                var topic = _topicService.GetTopicBySystemName(systemName, _storeContext.CurrentStore.Id);
-                if (topic == null)
-                    return null;
-                if (!topic.Published)
-                    return null;
-                //Store mapping
-                if (!_storeMappingService.Authorize(topic))
-                    return null;
-                //ACL (access control list)
-                if (!_aclService.Authorize(topic))
-                    return null;
-                return PrepareTopicModel(topic);
-            });
-
-            if (cacheModel == null)
+            var model = _topicModelFactory.PrepareTopicModelBySystemName(systemName);
+            if (model == null)
                 return Content("");
 
-            return PartialView(cacheModel);
+            return PartialView(model);
         }
 
         [HttpPost, ValidateInput(false)]
