+++ src/Presentation/Nop.Web/Factories/TopicModelFactory.cs
@@ -1,0 +1,147 @@
+﻿using System;
+using System.Linq;
+using Nop.Core;
+using Nop.Core.Caching;
+using Nop.Core.Domain.Topics;
+using Nop.Services.Customers;
+using Nop.Services.Localization;
+using Nop.Services.Security;
+using Nop.Services.Seo;
+using Nop.Services.Stores;
+using Nop.Services.Topics;
+using Nop.Web.Infrastructure.Cache;
+using Nop.Web.Models.Topics;
+
+namespace Nop.Web.Factories
+{
+    public partial class TopicModelFactory : ITopicModelFactory
+    {
+        #region Fields
+
+        private readonly ITopicService _topicService;
+        private readonly IWorkContext _workContext;
+        private readonly IStoreContext _storeContext;
+        private readonly ICacheManager _cacheManager;
+        private readonly IStoreMappingService _storeMappingService;
+        private readonly IAclService _aclService;
+        private readonly ITopicTemplateService _topicTemplateService;
+
+        #endregion
+
+        #region Constructors
+
+        public TopicModelFactory(ITopicService topicService,
+            IWorkContext workContext,
+            IStoreContext storeContext,
+            ICacheManager cacheManager,
+            IStoreMappingService storeMappingService,
+            IAclService aclService,
+            ITopicTemplateService topicTemplateService)
+        {
+            this._topicService = topicService;
+            this._workContext = workContext;
+            this._storeContext = storeContext;
+            this._cacheManager = cacheManager;
+            this._storeMappingService = storeMappingService;
+            this._aclService = aclService;
+            this._topicTemplateService = topicTemplateService;
+        }
+
+        #endregion
+
+        #region Utilities
+        
+        protected virtual TopicModel PrepareTopicModel(Topic topic)
+        {
+            if (topic == null)
+                throw new ArgumentNullException("topic");
+
+            var model = new TopicModel
+            {
+                Id = topic.Id,
+                SystemName = topic.SystemName,
+                IncludeInSitemap = topic.IncludeInSitemap,
+                IsPasswordProtected = topic.IsPasswordProtected,
+                Title = topic.IsPasswordProtected ? "" : topic.GetLocalized(x => x.Title),
+                Body = topic.IsPasswordProtected ? "" : topic.GetLocalized(x => x.Body),
+                MetaKeywords = topic.GetLocalized(x => x.MetaKeywords),
+                MetaDescription = topic.GetLocalized(x => x.MetaDescription),
+                MetaTitle = topic.GetLocalized(x => x.MetaTitle),
+                SeName = topic.GetSeName(),
+                TopicTemplateId = topic.TopicTemplateId
+            };
+            return model;
+        }
+        
+        #endregion
+
+        #region Methods
+
+        public virtual TopicModel PrepareTopicModelById(int topicId)
+        {
+            var cacheKey = string.Format(ModelCacheEventConsumer.TOPIC_MODEL_BY_ID_KEY,
+                topicId,
+                _workContext.WorkingLanguage.Id,
+                _storeContext.CurrentStore.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
+            var cacheModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var topic = _topicService.GetTopicById(topicId);
+                if (topic == null)
+                    return null;
+                if (!topic.Published)
+                    return null;
+                //Store mapping
+                if (!_storeMappingService.Authorize(topic))
+                    return null;
+                //ACL (access control list)
+                if (!_aclService.Authorize(topic))
+                    return null;
+                return PrepareTopicModel(topic);
+            });
+
+            return cacheModel;
+        }
+
+        public virtual TopicModel PrepareTopicModelBySystemName(string systemName)
+        {
+            var cacheKey = string.Format(ModelCacheEventConsumer.TOPIC_MODEL_BY_SYSTEMNAME_KEY,
+                systemName,
+                _workContext.WorkingLanguage.Id,
+                _storeContext.CurrentStore.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
+            var cacheModel = _cacheManager.Get(cacheKey, () =>
+            {
+                //load by store
+                var topic = _topicService.GetTopicBySystemName(systemName, _storeContext.CurrentStore.Id);
+                if (topic == null)
+                    return null;
+                if (!topic.Published)
+                    return null;
+                //ACL (access control list)
+                if (!_aclService.Authorize(topic))
+                    return null;
+                return PrepareTopicModel(topic);
+            });
+
+            return cacheModel;
+        }
+
+        public virtual string PrepareTemplateViewPath(int topicTemplateId)
+        {
+            var templateCacheKey = string.Format(ModelCacheEventConsumer.TOPIC_TEMPLATE_MODEL_KEY, topicTemplateId);
+            var templateViewPath = _cacheManager.Get(templateCacheKey, () =>
+            {
+                var template = _topicTemplateService.GetTopicTemplateById(topicTemplateId);
+                if (template == null)
+                    template = _topicTemplateService.GetAllTopicTemplates().FirstOrDefault();
+                if (template == null)
+                    throw new Exception("No default template could be loaded");
+                return template.ViewPath;
+            });
+            return templateViewPath;
+        }
+
+        #endregion
+    }
+}
