+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -288,6 +288,8 @@
     <Compile Include="Factories\ReturnRequestModelFactory.cs" />
     <Compile Include="Factories\IWidgetModelFactory.cs" />
     <Compile Include="Factories\IVendorModelFactory.cs" />
+    <Compile Include="Factories\ITopicModelFactory.cs" />
+    <Compile Include="Factories\TopicModelFactory.cs" />
     <Compile Include="Factories\VendorModelFactory.cs" />
     <Compile Include="Factories\WidgetModelFactory.cs" />
     <Compile Include="Global.asax.cs">
